% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Iran}
\alias{Iran}
\title{Shape file of for Iran (level 0)}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1 rows and 3 columns.
}
\source{
\url{https://gadm.org/}
}
\usage{
Iran
}
\description{
A dataset with geographic information about the polygons of Iran which is created by using read_sf("gadm36_IRN_0.shp") file form
https://gadm.org/
}
\keyword{datasets}
