% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_2.R
\name{team_2}
\alias{team_2}
\title{Generating the data frame associated with a shape file}
\usage{
team_2(file, tolerance)
}
\arguments{
\item{file}{Either a file path to the shape file or an data of class sf.}

\item{tolerance}{The value for thinning the polygon (default=0.1).}
}
\value{
It returns a data frame with longitude, latitude, order and group for the polygon of the given shape
file.
}
\description{
It takes either a file_path indicating the path to the shape file the shape file of class sf itself, and
creates a tidy data frame with lat, long, order and group attributes.
}
\examples{
data("Iran")
df <- team_2(Iran, 0.1)

}
